chore(docusaurus): split config for easy extending
#254
