+++ config/docusaurus/consts.js
@@ -1,0 +1,11 @@
+module.exports = {
+    DOMAIN: "https://feature-sliced.design/",
+    GITHUB_ORG: "https://github.com/feature-sliced",
+    GITHUB_DOCS: "https://github.com/feature-sliced/documentation",
+    TELEGRAM: "https://t.me/feature_sliced",
+    DISCORD: "https://discord.gg/S8MzWTUsmp",
+    YOUTUBE: "https://www.youtube.com/c/FeatureSlicedDesign",
+    TWITTER: "https://twitter.com/feature_sliced",
+    OPEN_COLLECTIVE: "https://opencollective.com/feature-sliced",
+    DEFAULT_LOCALE: "en",
+};
