+++ config/docusaurus/sidebars.community.js
